requirements.txt
PyQt5==5.15.6
PyOpenGL==3.1.6
numpy==1.24.3
